+++ packages/amplify-graphql-transformer-core/src/utils/rds-util.ts
@@ -67,5 +67,9 @@ export const constructFieldMappingInput = (): Expression => {
   ]);
 };
 
-export const constructAuthFilterStatement = (keyName: string): Expression =>
-  iff(not(methodCall(ref('util.isNullOrEmpty'), ref('ctx.stash.authFilter'))), set(ref(keyName), ref('ctx.stash.authFilter')));
+export const constructAuthFilterStatement = (keyName: string, emptyAuthFilter: boolean = false): Expression => {
+  if (emptyAuthFilter) {
+    return set(ref(keyName), obj({}));
+  }
+  return iff(not(methodCall(ref('util.isNullOrEmpty'), ref('ctx.stash.authFilter'))), set(ref(keyName), ref('ctx.stash.authFilter')));
+};